requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector==2.2.9
numpy==1.22.3
numpyencoder==0.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
